+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 	Changelog
 </h1>
 
+### 1.0.2
+
+#### New features
+
+-   output models
+-   game duration and players to getLiveGameData
+
+#### Changes
+
+-   separate router logic
+
 ### 1.0.1
 
 #### New features
@@ -15,4 +26,4 @@
 -   removal of base URL
 -   library function groups
 
-## 1.0.0
+# 1.0.0